+++ indico/modules/rb/controllers/backend/rooms.py
@@ -72,7 +72,7 @@ class RHSearchRooms(RHRoomBookingBase):
         rooms = [(id_, room_name) for id_, room_name, in search_query.with_entities(Room.id, Room.full_name)]
 
         # We can't filter by blocking's acl in the search_query, so we need to adjust the results
-        rooms = self._adjust_blockings(rooms, args, availability)
+        rooms = self._adjust_blockings(rooms, args, availability, admin_override_enabled)
         room_ids = [room[0] for room in rooms]
         if filter_availability:
             room_ids_without_availability_filter = [
@@ -91,11 +91,12 @@ class RHSearchRooms(RHRoomBookingBase):
             return None
         return [dt.date().isoformat() for dt in ReservationOccurrence.iter_start_time(start_dt, end_dt, repetition)]
 
-    def _adjust_blockings(self, rooms, filters, availability):
+    def _adjust_blockings(self, rooms, filters, availability, admin_override_enabled):
         if availability is None:
             return rooms
         blocked_rooms = get_blockings_with_rooms(filters['start_dt'].date(), filters['end_dt'].date())
-        nonoverridable_blocked_rooms = filter_blocked_rooms(blocked_rooms, nonoverridable_only=True)
+        nonoverridable_blocked_rooms = filter_blocked_rooms(blocked_rooms, nonoverridable_only=True,
+                                                            allow_admin=admin_override_enabled)
         if availability:
             # Remove nonoverridable blockings from available rooms
             nonoverridable_blocked_rooms_ids = [room.room_id for room in nonoverridable_blocked_rooms]